Use offsets instead of pointers everywhere
There are a lot of places where we need to know span and data offsets.
Recovering them from span pointers is tedious.

This patch replaces data pointer with data offset in pmemstream_span_runtime.
This allows us to use offsets everywhere and convert to pointers only when
needed (for memcpy or for reading the actual data). It makes the code a
little less verbose.