Always validate entry in iterator and check entry size
To support concurrent iteration and append we need to allow iterators to
only read fully written data. We could do that by having a runtime
'commited_offset' but it would be problematic for multi-process support.

Instead, always validate each entry - if it's size or popcount is invalid
it means that we either encountered a corrupted entry or have a race with append.
In both cases, return -1 to the user.