Define return type for make_pmemstream
Define it to be std::unique_ptr with std::function deleter.
Knowing type of handle returned by make_pmemstream makes it
easier to handle (e.g. to store as member in a class).

To make it work it's necessary to wrap stream_delete function
into a copyable type. This is done by make_copyable function:
it wraps any callable type into shared_ptr (which is copyable)
and returns function which dereferences the shared_ptr and
calls original function.