Define return type for make_pmemstream
Define it to be std::unique_ptr with std::function deleter.
Knowing type of handle returned by make_pmemstream makes it
easier to handle (e.g. to store as member in a class).

Deleter of map is stored in shared_ptr for simplicity
of handling its life time when it's captured in lambda expression.

Change is based on work of Igor Chorążewicz <igor.chorazewicz@intel.com>