Implement concurrent append
Use mpmc_queue to keep track of 'append' and 'commited' offsets.
During append, in reservae stage we acquire offset from the
mpmc_queue. At this point, the memory is exclusibvely owned by
the calling thread (other threads can still call acquire on
mpmc_queue to get ownership of memory with higher offset).

In publish, just before returning from function we publish
the offset which makes it available for consuming. Now,
we consume all offsets which are lower that end offset of newly
added entry. In case of multithreaded application this might
require spinning and waiting for other threads to call produce.