set initial persisted timestamp as INVALID_TIMESTAMP + 1
Previously we set persisted timestamp, on pmemstream start, to 0, but that
might have been misleading. In some cases '0' may be set improperly as a
timestamp and wrong entry could have been counted as a proper entry.
So now, in the recovery, there's a check to verify if currently checked entry
isn't timestamped with 'INVALID_TIMESTAMP' value.